refactor filters
Originally marked as october filter, it should have been a november
filter. Mongo images are dated +1 month comparing to openlibrary dumps
(which are released on the last day of the month). So Oct 2020 dump
would be a November 2020 mongo image.

Will rerun september dumps to recreate an october mongo image, and so
October filter will be added later.